Move view data traits from kas::widget::view::data_traits to kas::data
