+++ kas-wgpu/src/window.rs
@@ -10,11 +10,11 @@ use std::rc::Rc;
 use std::time::Instant;
 
 use kas::conv::Cast;
+use kas::data::SharedData;
 use kas::draw::{SizeHandle, ThemeAction, ThemeApi};
 use kas::event::{CursorIcon, ManagerState, UpdateHandle};
 use kas::geom::{Coord, Rect, Size};
 use kas::layout::SolveCache;
-use kas::widget::view::SharedData;
 use kas::{TkAction, WindowId};
 use kas_theme::Theme;
 use winit::dpi::PhysicalSize;
