+++ src/lib.rs
@@ -38,6 +38,7 @@ mod toolkit;
 // public implementations:
 pub mod class;
 pub mod conv;
+pub mod data;
 pub mod dir;
 pub mod draw;
 pub mod event;
