+++ src/widget/view/mod.rs
@@ -9,15 +9,7 @@
 //!
 //! # Shared data
 //!
-//! Shared data must implement one of a family of traits:
-//!
-//! -   [`SingleData`] supports viewing a single item ("datum")
-//! -   [`ListData`] supports viewing an ordered sequence of items (requires
-//!     that items are ordered and that items can be accessed by some key type)
-//!
-//! Each of these has a "Mut" variant, supporting direct modification of values
-//! when a mutable reference is available. In other cases, the `update` method
-//! *may* support modification.
+//! Shared data must implement one or more of the traits from [`kas::data`].
 //!
 //! ## Filters
 //!
@@ -37,16 +29,15 @@
 //!     Performance is potentially bounded by O(v) in all operations where `v`
 //!     is the number of visible items (depending on the [`ListData`] object).
 
-mod data_traits;
+#[allow(unused)]
+use kas::data::{ListData, SingleData};
+
 mod filter;
 mod list_view;
 mod shared_data;
 mod single_view;
 mod view_widget;
 
-pub use data_traits::{
-    ListData, ListDataMut, SharedData, SharedDataRec, SingleData, SingleDataMut,
-};
 pub use filter::{Filter, FilteredList, SimpleCaseInsensitiveFilter};
 pub use list_view::{ListMsg, ListView, SelectionMode};
 pub use shared_data::SharedRc;
